examples/all-in-one/requirements.txt
apache-airflow-providers-docker>=3.7,<4
apache-airflow-providers-postgres>=5,<6
typer>=0.12,<0.13
